refactor shared helper
